Improving the stampit code to be more readable
Cool examples!

I find it more clear when we declare dependencies explicitly.

For example `mover` is using `this.x` and `this.y`. This means it is better to clearly declare that `mover` depends on `haveCoordinates` stamp.
Same thing `slasher` is using `this.name`, thus it's better to emphasise that `slasher` inherits `haveName`.
Thus I split the `character` onto thee different parts - `haveCoordinates`, `slasher` and the additional props.lifePoints.